add parent class from WS tests
