+++ src/test/java/online/talkandtravel/controller/websocket/integrationtest/EventControllerIntegrationTest.java
@@ -1,0 +1,5 @@
+package online.talkandtravel.controller.websocket.integrationtest;
+
+public class EventControllerIntegrationTest {
+
+}
